refactor(batch): extract batch and IFC logic into lib modules
Move document opening, IFC export, model-list parsing and shared
reporting helpers into lib/tools/batch and a sibling ifc_ui module,
leaving script.py as a thin controller that only handles user
interaction and orchestration. Sets the ground for sharing these
helpers with the upcoming BatchMergeIFC command.
