+++ lib/tools/batch/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+"""Reusable services for pyArchitect batch commands."""
